done photo routes, added setters to models

+++ controllers/photoController.go
@@ -23,6 +23,9 @@ func CreatePhoto(c *gin.Context) {
 		return
 	}
 
+	authPayload := c.Value("User").(*helpers.AuthPayload)
+	photo.SetUserId(authPayload.UserId)
+
 	savedPhoto, err := models.CreatePhoto(&photo)
 
 	if err != nil {
@@ -47,6 +50,7 @@ func GetPhoto(c *gin.Context) {
 
 func UpdatePhoto(c *gin.Context) {
 	idStr, ok := c.Params.Get("id")
+
 	if !ok {
 		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": "Photo id not provided"})
 		return
@@ -66,7 +70,28 @@ func UpdatePhoto(c *gin.Context) {
 		return
 	}
 
-	c.JSON(http.StatusOK, photo)
+	authPayload := c.Value("User").(*helpers.AuthPayload)
+
+	if authPayload.UserId != photo.UserId {
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized wrong user"})
+		return
+	}
+
+	var photoUpdate models.Photo
+
+	if err = c.ShouldBindJSON(&photoUpdate); err != nil {
+		c.AbortWithStatusJSON(http.StatusBadRequest, gin.H{"error": err.Error()})
+		return
+	}
+
+	updatedPhoto, err := models.UpdatePhoto(uint(id), &photoUpdate)
+
+	if err != nil {
+		c.AbortWithStatusJSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
+		return
+	}
+
+	c.JSON(http.StatusOK, updatedPhoto)
 }
 
 func DeletePhoto(c *gin.Context) {
@@ -84,6 +109,20 @@ func DeletePhoto(c *gin.Context) {
 		return
 	}
 
+	photo, err := models.FindPhotoById(uint(id))
+
+	if err != nil {
+		c.AbortWithStatusJSON(http.StatusNotFound, gin.H{"error": err.Error()})
+		return
+	}
+
+	authPayload := c.Value("User").(*helpers.AuthPayload)
+
+	if authPayload.UserId != photo.UserId {
+		c.AbortWithStatusJSON(http.StatusUnauthorized, gin.H{"error": "unauthorized wrong user"})
+		return
+	}
+
 	err = models.DeletePhoto(uint(id))
 	
 	if err != nil {
@@ -92,6 +131,4 @@ func DeletePhoto(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{"deleted": true})
-
-
 }
